src/requirements.txt
faiss==1.9.0
numpy==1.22.4
opencv_python==4.10.0.84
Pillow==10.0.1
Pillow==11.1.0
scikit_learn==1.3.0
torch==2.2.0
torchvision==0.17.0
tk
pickle
